refactor(wizard): #1544 stage 7 — extract create_course enrollAndCreateCaller
Stage 7 of 8. Lifts the ~120-LOC enrollment block (TEACHER caller +
two test callers + cohort + memberships) from create_course.ts into
`tools/create_course/_enroll.ts` exporting `enrollAndCreateCaller(ctx)`.

Returns an `EnrollState` bundle the orchestrator hands to Stage 8 for
the return payload assembly: teacherCallerId, caller, callerName,
demoCaller, demoName, cohort, joinToken.

Preserves the 2026-05-19 cohort-by-playbookId scoping (the live fix
that prevented brand-new playbooks from inheriting member lists from
sibling courses with the same name) and the demo-caller best-effort
try/catch.

Orchestrator's `enrollCaller` + `randomFakeName` dynamic imports
removed — both moved into the helper.

Net file shape:
- create_course.ts: 376 → 255 LOC (-121)
- new: _enroll.ts (158 LOC)

Tests
- 12/12 dispatcher pins green
- 161/161 tests/lib/chat suite green
- 0 new tsc errors introduced
- AC respected: _enroll.ts well under the 400-LOC ceiling

Refs #1544
